server/app/requirements.txt
django==3.2.3
django-cors-headers==3.4.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
drf-yasg==1.20.0
Pillow==8.2.0
psycopg2-binary==2.8.6
gunicorn==20.0.4